+++ src/main/java/kr/merutilm/rff/approx/MPATable.java
@@ -2,7 +2,6 @@ package kr.merutilm.rff.approx;
 
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 
 import kr.merutilm.rff.formula.MandelbrotReference;
@@ -21,8 +20,8 @@ public abstract class MPATable {
     protected static final int REQUIRED_PERTURBATION = 2;
 
     protected MPATable(MandelbrotReference reference, MPASettings mpaSettings) {
-        int[] referencePeriod = reference.period();
-        int longestPeriod = reference.longestPeriod();
+        long[] referencePeriod = reference.period();
+        long longestPeriod = reference.longestPeriod();
         int minSkip = mpaSettings.minSkipReference();
 
         if (longestPeriod < minSkip) {
@@ -48,28 +47,28 @@ public abstract class MPATable {
 
         List<ArrayCompressionTool> refCompTools = refCompressor.tools();
         List<ArrayCompressionTool> mpaTools = new ArrayList<>();
-        int[] tablePeriod = mpaPeriod.tablePeriod;
-        int[] tablePeriodElements = mpaPeriod.tableElements;
+        long[] tablePeriod = mpaPeriod.tablePeriod;
+        long[] tablePeriodElements = mpaPeriod.tableElements;
         boolean[] isArtificial = mpaPeriod.isArtificial;
 
         for (ArrayCompressionTool tool : refCompTools) {
-            int start = tool.start();
-            int length = tool.range();
+            long start = tool.start();
+            long length = tool.range();
             int index = Arrays.binarySearch(tablePeriod, length + 1);
-            int tableIndex = iterationToPulledTableIndex(mpaPeriod, start);
+            long tableIndex = iterationToPulledTableIndex(mpaPeriod, start);
 
             // Check if the reference compressor is same as period.
             // However, The Computer doesn't know whether the compressor's length came from skipping to the periodic point, or being cut off in the middle.
             // So, Do check tableIndex too.
             if (index >= 0 && tableIndex >= 0 && !isArtificial[index]) {
-                int periodElements = tablePeriodElements[index];
+                long periodElements = tablePeriodElements[index];
                 mpaTools.add(new ArrayCompressionTool(1, tableIndex + 1, tableIndex + periodElements - 1));
             }
         }
         return new ArrayCompressor(mpaTools);
     }
 
-    protected static int iterationToPulledTableIndex(Period mpaPeriod, int iteration) {
+    protected static long iterationToPulledTableIndex(Period mpaPeriod, long iteration) {
         //
         // get index <=> Inverse calculation of index compression
         // First approach : check the remainder == 1
@@ -103,11 +102,11 @@ public abstract class MPATable {
             return -1;
         }
 
-        int[] tablePeriod = mpaPeriod.tablePeriod;
-        int[] tablePeriodElements = mpaPeriod.tableElements;
+        long[] tablePeriod = mpaPeriod.tablePeriod;
+        long[] tablePeriodElements = mpaPeriod.tableElements;
 
-        int index = 0;
-        int remainder = iteration;
+        long index = 0;
+        long remainder = iteration;
 
         for (int i = tablePeriod.length - 1; i >= 0; i--) {
             if (remainder < tablePeriod[i]) {
@@ -137,13 +136,13 @@ public abstract class MPATable {
     }
 
 
-    protected int iterationToCompTableIndex(int iteration) {
+    protected int iterationToCompTableIndex(long iteration) {
 
         return switch (settings.mpaCompressionMethod()) {
-            case NO_COMPRESSION -> iteration;
-            case LITTLE_COMPRESSION -> iterationToPulledTableIndex(mpaPeriod, iteration);
+            case NO_COMPRESSION -> (int)iteration;
+            case LITTLE_COMPRESSION -> (int)iterationToPulledTableIndex(mpaPeriod, iteration);
             case STRONGEST -> {
-                int index = iterationToPulledTableIndex(mpaPeriod, iteration);
+                long index = iterationToPulledTableIndex(mpaPeriod, iteration);
                 yield index == -1 ? -1 : pulledMPACompressor.compress(index);
             }
         };
@@ -152,10 +151,10 @@ public abstract class MPATable {
 
 
 
-    protected record Period(int[] tablePeriod, boolean[] isArtificial, int[] tableElements) {
+    protected record Period(long[] tablePeriod, boolean[] isArtificial, long[] tableElements) {
         @Override
         public boolean equals(Object o) {
-            return o instanceof Period(int[] t, boolean[] a, int[] te) &&
+            return o instanceof Period(long[] t, boolean[] a, long[] te) &&
                    Arrays.equals(tablePeriod, t) &&
                    Arrays.equals(isArtificial, a) &&
                    Arrays.equals(tableElements, te);
@@ -173,7 +172,7 @@ public abstract class MPATable {
             return Arrays.hashCode(tablePeriod) + Arrays.hashCode(tableElements);
         }
 
-        private static int[] generatePeriodElements(int[] tablePeriod) {
+        private static long[] generatePeriodElements(long[] tablePeriod) {
 
 
             // index compression : [3, 11, 26, 77] // index compression : [3, 11, 26, 77]
@@ -192,16 +191,16 @@ public abstract class MPATable {
             // period 77 : 77/26=2.xxx (7*2 elements), 77%26 = 25,25/11= 2.xxx (3*2 elements), 25%11=4, 4/3 = 1.xxx (1 element), elements = 7*2+3*2+1=21
             // Stored elements to memory
 
-            int[] tablePeriodElements = new int[tablePeriod.length];
+            long[] tablePeriodElements = new long[tablePeriod.length];
             for (int i = 0; i < tablePeriodElements.length; i++) {
                 if (i == 0) {
                     tablePeriodElements[i] = 1;
                     continue;
                 }
-                int elements = 0;
-                int remainder = tablePeriod[i];
+                long elements = 0;
+                long remainder = tablePeriod[i];
                 for (int j = i - 1; j >= 0; j--) {
-                    int groupAmount = remainder / tablePeriod[j];
+                    long groupAmount = remainder / tablePeriod[j];
                     remainder %= tablePeriod[j];
                     elements += groupAmount * tablePeriodElements[j];
                 }
@@ -210,7 +209,7 @@ public abstract class MPATable {
             return tablePeriodElements;
         }
 
-        private static Temp generateTablePeriod(int[] referencePeriod, MPASettings MPASettings) {
+        private static Temp generateTablePeriod(long[] referencePeriod, MPASettings MPASettings) {
 
             // example 
             // period : [3, 11, 26]
@@ -237,13 +236,13 @@ public abstract class MPATable {
 
             int maxMultiplier = MPASettings.maxMultiplierBetweenLevel();
             int minSkip = MPASettings.minSkipReference();
-            int longestPeriod = referencePeriod[referencePeriod.length - 1];
+            long longestPeriod = referencePeriod[referencePeriod.length - 1];
 
-            int[] tablePeriodArrayTemp = new int[1];
+            long[] tablePeriodArrayTemp = new long[1];
             boolean[] isArtificialArrayTemp = new boolean[1];
 
             int periodArraySize = 1;
-            int currentRefPeriod = minSkip;
+            long currentRefPeriod = minSkip;
 
 
             tablePeriodArrayTemp[0] = currentRefPeriod;
@@ -251,7 +250,7 @@ public abstract class MPATable {
             //first period is always minimum skip iteration when the longest period is larger than this,
             //and it is artificially-created period if generated period is not an element of generated period.
 
-            for (int p : referencePeriod) {
+            for (long p : referencePeriod) {
 
                 //Generate Period Array
 
@@ -290,22 +289,22 @@ public abstract class MPATable {
             }
 
 
-            int[] tablePeriod = Arrays.copyOfRange(tablePeriodArrayTemp, 0, periodArraySize);
+            long[] tablePeriod = Arrays.copyOfRange(tablePeriodArrayTemp, 0, periodArraySize);
             boolean[] isArtificial = Arrays.copyOfRange(isArtificialArrayTemp, 0, periodArraySize);
             return new Temp(tablePeriod, isArtificial);
         }
 
-        private static Period create(int[] referencePeriod, MPASettings MPASettings) {
+        private static Period create(long[] referencePeriod, MPASettings MPASettings) {
 
             Temp temp = generateTablePeriod(referencePeriod, MPASettings);
-            int[] tablePeriod = temp.tablePeriod();
+            long[] tablePeriod = temp.tablePeriod();
             boolean[] isArtificial = temp.isArtificial();
-            int[] tablePeriodElements = generatePeriodElements(tablePeriod);
+            long[] tablePeriodElements = generatePeriodElements(tablePeriod);
 
             return new Period(tablePeriod, isArtificial, tablePeriodElements);
         }
 
-        private record Temp(int[] tablePeriod, boolean[] isArtificial) {
+        private record Temp(long[] tablePeriod, boolean[] isArtificial) {
             @Override
             public String toString() {
                 return "";
