+++ src/main/java/kr/merutilm/rff/approx/PA.java
@@ -1,5 +1,5 @@
 package kr.merutilm.rff.approx;
 
 public interface PA {
-    int skip();
+    long skip();
 }
