+++ src/main/java/kr/merutilm/rff/approx/PABuilder.java
@@ -2,8 +2,8 @@ package kr.merutilm.rff.approx;
 
 public interface PABuilder<P extends PA> {
 
-    int start();
-    int skip();
+    long start();
+    long skip();
     PABuilder<P> step();
     PABuilder<P> merge(P toMerge);
     P build();
